requirements.txt
Django~=3.1.3
django-formwizard~=1.0
django-formtools~=2.2
numpy~=1.19.4
matplotlib~=3.3.3
mpld3~=0.5.2